requirements.txt
# Web Framework (updated for Python 3.13 compatibility)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# LLM Integration (install at least one)
openai>=1.40.0
# anthropic>=0.34.0  # Uncomment if using Anthropic

# Video Processing
ffmpeg-python>=0.2.0
opencv-python>=4.10.0
Pillow>=10.4.0

# HTTP Client
httpx>=0.27.0
requests>=2.32.0

# Task Queue (for async processing)
celery>=5.4.0
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.9

# Logging
loguru>=0.7.2

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

